+++ lib/nylas.rb
@@ -85,6 +85,8 @@ require_relative "nylas/thread"
 require_relative "nylas/webhook"
 require_relative "nylas/scheduler"
 require_relative "nylas/job_status"
+require_relative "nylas/token_info"
+require_relative "nylas/application_details"
 
 # Neural specific types
 require_relative "nylas/neural"